# openapi/v1.yaml
...
openapi: 3.1.0
info:
  title: cafaye darkroom
  version: 1.0.0
  summary: Media service — signed uploads, a tenant-scoped asset registry, and derived variants.
  description: |
    `darkroom` owns uploaded media for the platform. anychat's voice clips and
    anymark's attachments both sit on this API rather than each rolling their
    own.

    ## The upload flow

    Uploads are three calls, and the bytes never pass through this service:

    1. `POST /v1/uploads` with the file's name, type, size and sha256. The
       response is a **presigned PUT URL**, scoped to exactly one storage key,
       one content type and one maximum size, valid for **15 minutes**.
    2. The client `PUT`s the bytes **straight to object storage**. Not to this
       service. A 1 GiB upload is one request to the bucket, not a stream
       through an API process that has to buffer it.
    3. `POST /v1/uploads/{id}/complete`. The service reads the object's metadata,
       computes the sha256 of what storage actually holds, compares it to what
       the client claimed, and only then marks the asset `ready`.

    Step 3 is the security property, and it is the reason step 2 is direct. The
    checksum the client sent at step 1 is a *claim*. At step 3 it is checked
    against reality, and the row ends up carrying the computed value. A service
    that trusted the claim would let a client register an asset whose checksum
    describes bytes it never uploaded — and that checksum is what every
...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: [assets:read, assets:write]
paths:
...
  /v1/assets:
    get:
      tags: [assets]
      operationId: listAssets
      summary: List the caller's account's assets
      description: |
        Cursor-paginated and scoped to the `account_id` in the caller's token.
        An asset belonging to another account is simply not in the result, and
        the total count does not include it either.
      parameters:
        - name: limit
          in: query
          description: Defaults to 25, capped at 100. A larger request is clamped, not rejected.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
        - name: cursor
          in: query
          description: Opaque base64url. Clients must not parse it.
          schema:
            type: string
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: One page. `data` is always an array, empty rather than absent.
          headers:
            X-Trace-Id:
              $ref: '#/components/headers/TraceId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v1/assets/{id}:
    get:
      tags: [assets]
      operationId: getAsset
      summary: Fetch one asset
      description: |
        Scoped to the caller's account. An asset belonging to another account
        returns `404` with a body **identical** to the one returned for an id
        that never existed — a `403` would be an asset-id oracle.
      parameters:
        - $ref: '#/components/parameters/AssetId'
      responses:
        '200':
          description: The asset.
          headers:
            X-Trace-Id:
              $ref: '#/components/headers/TraceId'
          content:
            application/json:
              schema:
                type: object
                required: [asset]
                properties:
                  asset:
                    $ref: '#/components/schemas/Asset'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/Unavailable'
    delete:
      tags: [assets]
      operationId: deleteAsset
      summary: Delete an asset
      description: |
        Removes the row, removes the storage objects — the original **and every
        variant** — and writes `darkroom.asset.deleted` to the outbox in the
        same transaction.

        Objects are deleted before the row commits. A row that says "gone" while
        a gigabyte of objects survives is a leak nobody will ever find, whereas
        an object removed just before a rolled-back row is recoverable.

        Idempotent: a second delete is `404`, not a `500`.
      parameters:
        - $ref: '#/components/parameters/AssetId'
      responses:
        '204':
          description: Deleted. The storage objects are gone.
          headers:
            X-Trace-Id:
              $ref: '#/components/headers/TraceId'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/Unavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A token issued by identity. darkroom verifies it locally against
        identity's JWKS with a bounded cache; identity is never on the hot
        path.

        Required claims: `iss`, `aud`, `sub`, `exp`, `account_id`, `scope`.
        `roles` is **not** read — core's convention is that services check
        scopes, not roles.

        Accepted algorithms are RS256 and ES256. `alg: none`, HS256, and any
        algorithm the key does not advertise are rejected outright.

        `account_id` is the tenancy boundary and is taken from this token and
        from nowhere else. There is no account field in any request body.
      x-required-scopes:
        read: [assets:read]
        write: [assets:write]
  headers:
    TraceId:
      description: Always present, always equal to `trace_id` in the problem body.
      schema:
        type: string
...
  parameters:
    AssetId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Malformed request syntax.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: No credential, or one that did not verify.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: |
        Authenticated, but the token lacks the required scope. Never used for a
        resource under another account — that is a `404`.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: No such asset **in the caller's account**.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ValidationFailed:
      description: Semantically wrong. `errors[]` lists the per-field failures.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unavailable:
      description: A dependency is not answering. Safe to retry.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      description: |
        The cafaye error envelope: RFC 9457 plus core's `code` and `trace_id`.
        Every non-2xx response from this service is this shape.
      required: [type, title, status, detail, instance, code, trace_id]
      properties:
        type:
          type: string
          format: uri
          example: https://errors.cafaye.com/validation_failed
        title:
          type: string
          description: A fixed summary for the code. May be reworded without a version bump.
        status:
          type: integer
        detail:
          type: string
          description: |
            Specific to this occurrence, never parsed by clients. Never contains
            an internal error, a host, a role or a query fragment.
        instance:
          type: string
          description: The request path, never including the query string.
        code:
          type: string
          enum:
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - validation_failed
            - rate_limited
            - idempotency_key_reused
            - internal
            - unavailable
        trace_id:
          type: string
          description: Always present, always equal to the `X-Trace-Id` header.
        errors:
          type: array
          description: Present only on 422.
          items:
            type: object
            required: [field, code]
            properties:
              field:
                type: string
              code:
                type: string
...
    AssetKind:
      type: string
      enum: [image, audio, video, document]
      description: |
        Derived server-side from `content_type`, never client-supplied. A `kind`
        chooses the decoder, so a client that could set one could route a
        request through a decoder it should not reach.
    AssetStatus:
      type: string
      enum: [pending, ready, failed]
      description: |
        `pending` means a URL was issued and nothing is confirmed. `ready` and
        `failed` are both terminal.
    Asset:
      type: object
      required:
        - id
        - account_id
        - owner_user_id
        - kind
        - original_filename
        - content_type
        - byte_size
        - checksum
        - status
        - metadata
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
          description: Taken from the caller's token, never from the request.
        owner_user_id:
          type: string
          format: uuid
          description: '`sub` — the user who created the upload.'
        kind:
          $ref: '#/components/schemas/AssetKind'
        original_filename:
          type: string
          maxLength: 512
        content_type:
          type: string
          example: image/png
        byte_size:
          type: integer
          format: int64
          maximum: 1073741824
        checksum:
          type: string
          pattern: '^[0-9a-f]{64}$'
          description: |
            Lowercase sha256 of the bytes **as stored**. Not the client's claim —
            the claim is verified at complete time and replaced by the computed
            value.
        status:
          $ref: '#/components/schemas/AssetStatus'
        metadata:
          type: object
          additionalProperties: true
          description: |
            Free-form. On a `failed` asset it carries `failure_reason`:
            `object_absent_at_complete` or `checksum_mismatch`.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      description: |
        There is deliberately no `storage_key` field. The key is a server-side
        implementation detail of where the bytes live, and a client that can
        read it will eventually try to build a URL from it.
...
    Page:
      type: object
      required: [next_cursor, has_more]
      properties:
        next_cursor:
          type: [string, 'null']
          description: Opaque. Null on the last page.
        has_more:
          type: boolean
    AssetList:
      type: object
      required: [data, page]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Asset'
        page:
          $ref: '#/components/schemas/Page'
...